# docs/swagger.yaml
definitions:
  cms_internal_app_category_internal_dto.ApiPagedData-model_Category:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  dto.ApiArticle:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.Article'
      msg:
        type: string
    type: object
...
  dto.ApiCategory:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/dto.Category'
      msg:
        type: string
    type: object
  dto.ApiCategoryList:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/cms_internal_app_category_internal_dto.ApiPagedData-model_Category'
      msg:
        type: string
    type: object
...
  dto.Article:
    properties:
      authorId:
        type: string
      authorName:
        type: string
      categoryId:
        type: string
      categoryName:
        type: string
      clickCount:
        type: string
      content:
        type: string
      createdAt:
        type: string
      description:
        type: string
      flag:
        type: integer
      id:
        type: string
      keywords:
        items:
          type: string
        type: array
      modelData:
        additionalProperties: {}
        type: object
      modelId:
        type: string
      modelSchema:
        items:
          $ref: '#/definitions/dto.ArticleModelSchema'
        type: array
      publishAt:
        type: string
      status:
        type: integer
      target:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
  dto.ArticleCreateParams:
    properties:
      categoryId:
        type: string
      content:
        maxLength: 65535
        type: string
      description:
        maxLength: 255
        type: string
      flag:
        type: integer
      keywords:
        items:
          type: string
        maxItems: 10
        type: array
      modelData:
        additionalProperties: {}
        type: object
      modelID:
        type: string
      target:
        type: string
      title:
        maxLength: 255
        type: string
      url:
        type: string
    required:
    - categoryId
    - title
    type: object
...
  dto.ArticleModelSchema:
    properties:
      fieldKey:
        type: string
      fieldName:
        type: string
      sequence:
        type: string
      type:
        format: int32
        type: integer
    type: object
...
  dto.ArticleUpdateParams:
    properties:
      categoryId:
        type: string
      content:
        type: string
      description:
        type: string
      flag:
        type: integer
      id:
        type: string
      keywords:
        items:
          type: string
        type: array
      modelData:
        additionalProperties: {}
        type: object
      target:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  dto.Category:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      seo:
        $ref: '#/definitions/dto.CategorySEO'
      sequence:
        type: integer
      type:
        type: integer
      updatedAt:
        type: string
      visible:
        type: boolean
    type: object
  dto.CategoryCreateParams:
    properties:
      name:
        type: string
      parentID:
        type: string
      path:
        type: string
      seo:
        $ref: '#/definitions/dto.CategorySEO'
      sequence:
        type: integer
      type:
        type: integer
      visible:
        type: boolean
    required:
    - name
    - path
    type: object
  dto.CategoryListParams:
    properties:
      id:
        type: string
      name:
        type: string
      pageNo:
        type: integer
      pageSize:
        type: integer
      parentId:
        type: string
      path:
        type: string
      type:
        type: integer
      visible:
        type: boolean
    type: object
  dto.CategoryMoveParams:
    properties:
      id:
        type: string
      targetId:
        type: string
    type: object
  dto.CategorySEO:
    properties:
      description:
        type: string
      keywords:
        type: string
      title:
        type: string
    type: object
  dto.CategoryUpdateParams:
    properties:
      id:
        type: string
      name:
        type: string
      path:
        type: string
      seo:
        $ref: '#/definitions/dto.CategorySEO'
      sequence:
        type: integer
      type:
        type: integer
      visible:
        type: boolean
    required:
    - id
    type: object
...
  model.Category:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      parentID:
        type: integer
      path:
        type: string
      seo:
        $ref: '#/definitions/model.CategorySeo'
      sequence:
        type: integer
      type:
        type: integer
      updatedAt:
        type: string
      visible:
        type: integer
    type: object
  model.CategorySeo:
    properties:
      categoryID:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      keywords:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  types.ApiCreateResult:
    properties:
      code:
        type: string
      data:
        description: 数据 ID
        example: "123456"
        type: string
      msg:
        type: string
    type: object
  types.ApiEmptyResult:
    properties:
      code:
        type: string
      msg:
        type: string
    type: object
  types.ResourceID:
    properties:
      id:
        example: "123456"
        type: string
    required:
    - id
    type: object
info:
  contact: {}
  license:
    name: Apache 2.0
  title: EgoCms Api
  version: "0.1"
paths:
...
  /article/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleCreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiCreateResult'
      security:
      - ApiKeyAuth: []
      summary: 创建文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 删除文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/detail:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiArticle'
      security:
      - ApiKeyAuth: []
      summary: 查看文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/offline:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 下线文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 发布文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 拒审文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/republish:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 重新提交审核文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 提交文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /article/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ArticleUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 更新文章
      tags:
      - 文章
      x-apifox-folder: 文章
  /category/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryCreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiCreateResult'
      security:
      - ApiKeyAuth: []
      summary: 创建分类
      tags:
      - 分类
      x-apifox-folder: 文章分类
  /category/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 删除分类
      tags:
      - 分类
      x-apifox-folder: 文章分类
  /category/detail:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.ResourceID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiCategory'
      security:
      - ApiKeyAuth: []
      summary: 查看分类
      tags:
      - 分类
      x-apifox-folder: 文章分类
  /category/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiCategoryList'
      security:
      - ApiKeyAuth: []
      summary: 分类列表
      tags:
      - 分类
      x-apifox-folder: 文章分类
  /category/move:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryMoveParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 移动分类
      tags:
      - 分类
      x-apifox-folder: 文章分类
  /category/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiEmptyResult'
      security:
      - ApiKeyAuth: []
      summary: 更新分类
      tags:
      - 分类
      x-apifox-folder: 文章分类
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"